Small compilation fix for IPX_TYPE
IPX_TYPE is not available, I'm not overly familiar with the IPX code history but this was used to change the type of IPX packets. In all cases, this would have been set to the default of 1 - and since the NetCI base doesn't override this, we are just setting it here. IPX is not tested, might be something to deprecate later on.